test: 共通テストに移動
ついでに共通テスト用にデータ作成ロジックを個人ごとに変えられるような機能を追加
