+++ backend-svc/src/ruby_backend_svc/src/yamamoto/main.rb
@@ -1,4 +1,4 @@
-module Kamada
+module Yamamoto
   def self.main(request_json)
     '{status: "ok"}'
   end
