+++ backend-svc/src/ruby_backend_svc/test/common/kamada/test_data_operator.rb
@@ -1,0 +1,11 @@
+module Kamada
+  class TestDataOperator
+    def self.create
+      puts "Kamada Data Create!!"
+    end
+
+    def self.clean
+      puts "Kamada Data Crean!!"
+    end
+  end
+end
