+++ backend-svc/src/ruby_backend_svc/test/common/yamamoto/test_data_operator.rb
@@ -1,0 +1,11 @@
+module Yamamoto
+  class TestDataOperator
+    def self.create
+      puts "Yamamoto Data Create!!"
+    end
+
+    def self.clean
+      puts "Yamamoto Data Crean!!"
+    end
+  end
+end
