+++ backend-svc/src/ruby_backend_svc/test/common/yoneoka/test_data_operator.rb
@@ -1,0 +1,11 @@
+module Yoneoka
+  class TestDataOperator
+    def self.create
+      puts "Yoneoka Data Create!!"
+    end
+
+    def self.clean
+      puts "Yoneoka Data Clean!!"
+    end
+  end
+end
